+++ src/db/strategies/base/contextStrategy.js
@@ -12,8 +12,8 @@ class ContextStrategy extends Icrud {
         return this._database.create(item)
     }
 
-    read(item) {
-        return this._database.read(item)
+    read(item, skip, limit) {
+        return this._database.read(item, skip, limit)
     }
 
     update(id, item) {
