+++ src/db/strategies/mongo/schemas/heroSchemaMongo.js
@@ -1,0 +1,18 @@
+const Mongoose = require('mongoose')
+
+const heroSchema = new Mongoose.Schema({
+    name: {
+        type: String,
+        required: true
+    },
+    power: {
+        type: String,
+        required: true
+    },
+    insertedAt: {
+        type: Date,
+        default: new Date()
+    }
+})
+
+module.exports = Mongoose.model('heroesmarvel', heroSchema)
